configure wanleg userspace
