--- keyboards/gherkin/keymaps/wanleg/config.h
@@ -1,24 +1,0 @@
-#ifndef CONFIG_USER_H
-#define CONFIG_USER_H
-
-#include "../../config.h"
-#endif
-
-#define PREVENT_STUCK_MODIFIERS
-
-//Tap Dance Prerequisite
-#define TAPPING_TERM 200
-
-//Mousekeys Settings
-#define MOUSEKEY_INTERVAL 16
-#define MOUSEKEY_DELAY 0
-#define MOUSEKEY_TIME_TO_MAX 60
-#define MOUSEKEY_MAX_SPEED 7
-#define MOUSEKEY_WHEEL_DELAY 0
-
-/* for QMK DFU bootloader */
-/* not required if using default ProMicro bootloader */
-/* set top left key as bootloader mode escape key */
-#define QMK_ESC_OUTPUT B4 // usually COL
-#define QMK_ESC_INPUT F7 // usually ROW
-#define QMK_LED B0
