+++ layouts/community/ortho_3x10/readme.md
@@ -1,0 +1,3 @@
+# ortho_3x10
+
+    LAYOUT_ortho_3x10
