+++ users/wanleg/config.h
@@ -1,0 +1,58 @@
+#ifndef USERSPACE_CONFIG_H
+#define USERSPACE_CONFIG_H
+
+#define PREVENT_STUCK_MODIFIERS
+
+// TAPPING_TERM
+#ifdef TAP_DANCE_ENABLE
+#ifdef TAPPING_TERM
+#undef TAPPING_TERM
+#endif 
+#define TAPPING_TERM 200
+#endif
+
+//Mousekey Settings
+#ifdef MOUSEKEY_INTERVAL
+#undef MOUSEKEY_INTERVAL
+#endif
+#define MOUSEKEY_INTERVAL 16
+
+#ifdef MOUSEKEY_DELAY
+#undef MOUSEKEY_DELAY
+#endif
+#define MOUSEKEY_DELAY 0
+
+#ifdef MOUSEKEY_TIME_TO_MAX
+#undef MOUSEKEY_TIME_TO_MAX
+#endif
+#define MOUSEKEY_TIME_TO_MAX 60
+
+#ifdef MOUSEKEY_MAX_SPEED
+#undef MOUSEKEY_MAX_SPEED
+#endif
+#define MOUSEKEY_MAX_SPEED 7
+
+#ifdef MOUSEKEY_WHEEL_DELAY
+#undef MOUSEKEY_WHEEL_DELAY
+#endif
+#define MOUSEKEY_WHEEL_DELAY 0
+
+// Disable action_get_macro and fn_actions, since we don't use these
+// and it saves on space in the firmware.
+#ifndef NO_DEBUG
+#define NO_DEBUG
+#endif // !NO_DEBUG
+#if !defined(NO_PRINT) && !defined(CONSOLE_ENABLE)
+#define NO_PRINT
+#endif // !NO_PRINT
+#define NO_ACTION_MACRO
+#define NO_ACTION_FUNCTION
+
+//********** TESTING ****************//
+//when trying to type the rolling combo "zx" and z is configured to send Ctrl when hold,
+//z rapidly followed by x actually sends Ctrl-x. That's bad.
+//You can disable this behavior by defining IGNORE_MOD_TAP_INTERRUPT in config.h.
+#define IGNORE_MOD_TAP_INTERRUPT
+//********** TESTING ****************//
+
+#endif // !USERSPACE_CONFIG_H
