+++ users/wanleg/rules.mk
@@ -1,6 +1,11 @@
-TAP_DANCE_ENABLE = yes		# Enable Tap Dance (comment if not being implemented)
+SRC += wanleg.c tapdances.c
 
-#If ProMicro has QMK DFU bootloader instead of Caterina,
+ifndef TAP_DANCE_ENABLE
+    TAP_DANCE_ENABLE = yes
+endif
+
+
+#If using a ProMicro and it has the QMK DFU bootloader instead of Caterina,
 #run "make <keyboard>:<keymap> dfu=qmk" when compiling to ensure it is flagged properly after being flashed
 ifeq ($(strip $(dfu)), qmk)
     BOOTLOADER = qmk-dfu
