+++ users/wanleg/tapdances.c
@@ -1,0 +1,110 @@
+//Tap Dance Settings
+#include "wanleg.h"
+
+///// QUAD FUNCTION TAP DANCE GENERAL SETUP SECTION START /////
+///// (no need to edit this section) /////
+//Enums used to clearly convey the state of the tap dance
+enum {
+  SINGLE_TAP = 1,
+  SINGLE_HOLD = 2,
+  DOUBLE_TAP = 3,
+  DOUBLE_HOLD = 4,
+  DOUBLE_SINGLE_TAP = 5 //send SINGLE_TAP twice - NOT DOUBLE_TAP
+  // Add more enums here if you want for triple, quadruple, etc.
+};
+
+typedef struct {
+  bool is_press_action;
+  int state;
+} tap;
+
+int cur_dance (qk_tap_dance_state_t *state) {
+  if (state->count == 1) {
+    //If count = 1, and it has been interrupted - it doesn't matter if it is pressed or not: Send SINGLE_TAP
+    if (state->interrupted || !state->pressed) return SINGLE_TAP;
+    if (state->interrupted) return SINGLE_TAP;
+    else return SINGLE_HOLD;
+  }
+  //If count = 2, and it has been interrupted - assume that user is trying to type the letter associated
+  //with single tap.
+  else if (state->count == 2) {
+    if (state->interrupted) return DOUBLE_SINGLE_TAP;
+    else if (state->pressed) return DOUBLE_HOLD;
+    else return DOUBLE_TAP;
+  }
+  else return 6; //magic number. At some point this method will expand to work for more presses
+}
+///// QUAD FUNCTION TAP DANCE GENERAL SETUP SECTION END /////
+///// QUAD FUNCTION TAP DANCE PERSONALIZATION SECTION START /////
+//instantialize an instance of 'tap' for the 'ENT' tap dance.
+static tap ENTtap_state = {
+  .is_press_action = true,
+  .state = 0
+};
+
+void ENT_finished (qk_tap_dance_state_t *state, void *user_data) {
+  ENTtap_state.state = cur_dance(state);
+  switch (ENTtap_state.state) {
+    case SINGLE_TAP: register_code(KC_SPC); break;
+    case SINGLE_HOLD: register_code(KC_LSFT); break;
+    case DOUBLE_TAP: register_code(KC_ENT); break;
+    case DOUBLE_HOLD: register_code(KC_NO); break; // setting double hold to do nothing (change this if you want)
+    case DOUBLE_SINGLE_TAP: register_code(KC_SPC); unregister_code(KC_SPC); register_code(KC_SPC);
+    //Last case is for fast typing. Assuming your key is `f`:
+    //For example, when typing the word `buffer`, and you want to make sure that you send `ff` and not `Esc`.
+    //In order to type `ff` when typing fast, the next character will have to be hit within the `TAPPING_TERM`, which by default is 200ms.
+  }
+}
+
+void ENT_reset (qk_tap_dance_state_t *state, void *user_data) {
+  switch (ENTtap_state.state) {
+    case SINGLE_TAP: unregister_code(KC_SPC); break;
+    case SINGLE_HOLD: unregister_code(KC_LSFT); break;
+    case DOUBLE_TAP: unregister_code(KC_ENT); break;
+    case DOUBLE_HOLD: unregister_code(KC_NO);
+    case DOUBLE_SINGLE_TAP: unregister_code(KC_SPC);
+  }
+  ENTtap_state.state = 0;
+}
+
+//instanalize an instance of 'tap' for the 'DEL' tap dance.
+static tap DELtap_state = {
+  .is_press_action = true,
+  .state = 0
+};
+
+void DEL_finished (qk_tap_dance_state_t *state, void *user_data) {
+  DELtap_state.state = cur_dance(state);
+  switch (DELtap_state.state) {
+    case SINGLE_TAP: register_code(KC_BSPC); break;
+    case SINGLE_HOLD: register_code(KC_LCTL); break;
+    case DOUBLE_TAP: register_code(KC_DEL); break;
+    case DOUBLE_HOLD: register_code(KC_NO); break;
+    case DOUBLE_SINGLE_TAP: register_code(KC_BSPC); unregister_code(KC_BSPC); register_code(KC_BSPC);
+  }
+}
+
+void DEL_reset (qk_tap_dance_state_t *state, void *user_data) {
+  switch (DELtap_state.state) {
+    case SINGLE_TAP: unregister_code(KC_BSPC); break;
+    case SINGLE_HOLD: unregister_code(KC_LCTL); break;
+    case DOUBLE_TAP: unregister_code(KC_DEL); break;
+    case DOUBLE_HOLD: unregister_code(KC_NO);
+    case DOUBLE_SINGLE_TAP: unregister_code(KC_BSPC);
+  }
+  DELtap_state.state = 0;
+}
+///// QUAD FUNCTION TAP DANCE PERSONALIZATION SECTION END /////
+
+//Tap Dance Definitions
+//THIS SECTION HAS TO BE AT THE END OF THE TAP DANCE SECTION
+qk_tap_dance_action_t tap_dance_actions[] = {
+  [TD_SFT_CAPS] = ACTION_TAP_DANCE_DOUBLE(KC_LSFT, KC_CAPS)
+// Other declarations would go here, separated by commas, if you have them
+ ,[TD_Q_ESC]  = ACTION_TAP_DANCE_DOUBLE(KC_Q, KC_ESC)
+ ,[ENT_TAP_DANCE] = ACTION_TAP_DANCE_FN_ADVANCED(NULL, ENT_finished, ENT_reset)
+ ,[DEL_TAP_DANCE] = ACTION_TAP_DANCE_FN_ADVANCED(NULL, DEL_finished, DEL_reset)
+};
+
+//In Layer declaration, add tap dance item in place of a key code
+//TD(TD_SFT_CAPS)
